+++ include/starsquatters.h
@@ -33,7 +33,7 @@
 
 #include <SDL_ttf.h>
 
-#define PI M_PI
+#define PI ((float)M_PI)
 
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
